[Refactor] Fix filename typo
