--- TestShell/CmdLancher.cpp
@@ -1,1 +1,0 @@
-#include "CmdLancher.h"
