+++ TestShell/TestShell.vcxproj.filters
@@ -33,13 +33,7 @@
     <ClCompile Include="test_shell.cpp">
       <Filter>테스트 파일</Filter>
     </ClCompile>
-    <ClCompile Include="CmdLancher.cpp">
-      <Filter>소스 파일</Filter>
-    </ClCompile>
-    <ClCompile Include="$(MSBuildThisFileDirectory)..\..\lib\native\src\gtest\gtest-all.cc">
-      <Filter>소스 파일</Filter>
-    </ClCompile>
-    <ClCompile Include="$(MSBuildThisFileDirectory)..\..\lib\native\src\gmock\gmock-all.cc">
+    <ClCompile Include="CmdLauncher.cpp">
       <Filter>소스 파일</Filter>
     </ClCompile>
     <ClCompile Include="test_logger.cpp">
@@ -65,7 +59,7 @@
     <ClInclude Include="TestScripts.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
-    <ClInclude Include="CmdLancher.h">
+    <ClInclude Include="CmdLauncher.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
     <ClInclude Include="Shell.h">
